Start work on AvailableApps
