+++ frontend/src/resources/AppRegistration.js
@@ -1,0 +1,5 @@
+export default {
+  dataModel: {
+    types: ['interop:ApplicationRegistration']
+  }
+};
