--- frontend/src/resources/AppRegistration/index.js
@@ -1,18 +1,0 @@
-import AppRegistrationList from './AppRegistrationList';
-
-export default {
-  config: {
-    list: AppRegistrationList
-  },
-  dataModel: {
-    types: ['interop:ApplicationRegistration']
-  },
-  translations: {
-    en: {
-      name: 'Application |||| Applications'
-    },
-    fr: {
-      name: 'Application |||| Applications'
-    }
-  }
-};
